+++ src/assets/styles/_branding.scss
@@ -7,6 +7,7 @@ $app-accent-color: #6297b1;
 $app-background-color: white;
 
 $ds-accent-color: #20154E;
+$ds-dusty-gray-color: #9A9A9A;
 
 $button-accent-color: $app-accent-color;
 $button-accent-text-color: white;
