+++ src/components/Feed/FeedDefaultItem/FeedDefaultItem.scss
@@ -1,3 +1,5 @@
+@import "../../../assets/styles/branding";
+
 .feed-item {
   margin: 15px auto;
   max-width: 500px;
@@ -35,16 +37,12 @@
     overflow: hidden;
     font-size: 1.3rem;
 
-    //::ng-deep .mat-card-header-text {
-    //  margin: 0;
-    //}
-
     .feed-card-header {
       padding: 5px 15px;
 
       .card-header-title {
         font-size: 1.6rem;
-        color: #4b566c;
+        color: $app-dark-tone;
         margin-top: 10px;
         margin-bottom: 8px;
       }
@@ -86,7 +84,7 @@
         margin: 20px 15px;
         padding: 0;
         font-size: 13px;
-        color: #9a9a9a;
+        color: $ds-dusty-gray-color;
         line-height: 1.54;
 
         .feed-item-content {
@@ -118,14 +116,14 @@
       margin: 0;
       padding: 0 15px 10px 15px;
       font-size: 1.3rem;
-      color: #9a9a9a;
+      color: $ds-dusty-gray-color;
       font-style: italic;
 
       .card-expand-tag {
         font-weight: bold;
         font-style: normal;
         cursor: pointer;
-        color: #6297b1;
+        color: $app-accent-color;
       }
     }
 
@@ -139,118 +137,7 @@
       font-style: italic;
       text-transform: capitalize;
       font-size: 13px;
-      color: #9a9a9a;
+      color: $ds-dusty-gray-color;
     }
   }
 }
-
-//.feed-card {
-//  flex: 1 1;
-//  //margin: 0 0 0 15px;
-//  //max-width: 450px;
-//  box-shadow: none;
-//  border-radius: 4px;
-//  padding: 0;
-//  overflow: hidden;
-//  background: #fff;
-//  color: rgba(0,0,0,.87);
-//
-//  .feed-card-header {
-//    display: flex;
-//    align-items: center;
-//    border-bottom: 1px solid #f6f6f6;
-//    padding: 5px 15px;
-//    flex-direction: row;
-//    margin-top: 0;
-//  }
-//
-//  .feed-card-content {
-//    display: block;
-//    margin: 20px 15px;
-//    padding: 0;
-//    font-size: 13px;
-//    color: #9a9a9a;
-//    line-height: 1.54;
-//    position: relative;
-//    text-align: left;
-//
-//    .feed-item-overlay {
-//      background-color: rgba(29,32,59, 0.54);
-//      padding: 20px 15px;
-//      position: absolute;
-//      top: 0;
-//      bottom: 0;
-//      left: 0;
-//      right: 0;
-//      display: flex;
-//      flex-direction: column;
-//      justify-content: space-between;
-//
-//      .feed-item-overlay-content {
-//        overflow-wrap: break-word;
-//        word-wrap: break-word;
-//        white-space: pre-wrap;
-//        color: whitesmoke;
-//        font-size: 13px;
-//        line-height: 1.54;
-//      }
-//
-//      .feed-item-overlay-footer {
-//        color: whitesmoke;
-//        font-size: 13px;
-//        font-style: italic;
-//      }
-//    }
-//
-//    &.no-image {
-//      margin: 20px 15px;
-//      padding: 0;
-//      font-size: 13px;
-//      color: #9a9a9a;
-//      line-height: 1.54;
-//
-//      .feed-item-content {
-//        overflow-wrap: break-word;
-//        word-wrap: break-word;
-//        white-space: pre-wrap;
-//        text-overflow: ellipsis;
-//        overflow: hidden;
-//        max-height: 40px;
-//        transition: max-height 1s ease-out;
-//
-//        &.expanded {
-//          max-height: 1000px;
-//        }
-//      }
-//    }
-//
-//    &.single-image {
-//      background: center center no-repeat;
-//      background-size: cover;
-//      padding: 0;
-//      min-height: 200px;
-//    }
-//  }
-//  .card-header-action {
-//    font-style: italic;
-//    text-transform: capitalize;
-//    font-size: 13px;
-//    color: #9a9a9a;
-//  }
-//
-//  .feed-card-footer {
-//    display: flex;
-//    margin: 0;
-//    padding: 0 15px 10px 15px;
-//    font-size: 13px;
-//    color: #9a9a9a;
-//    font-style: italic;
-//
-//    .card-expand-tag {
-//      font-weight: bold;
-//      font-style: normal;
-//      cursor: pointer;
-//      color: #6297b1;
-//    }
-//  }
-//}
