+++ src/components/Feed/WeeklySummary/WeeklySummary.scss
@@ -51,7 +51,6 @@
 .source-icon {
   float: left;
   text-align: center;
-  //margin-top: 9px;
   border-radius: 50%;
 }
 
@@ -79,7 +78,6 @@
   box-sizing: border-box;
   overflow: hidden;
   flex-direction: row;
-  //align-items: center;
   padding-top: 10px;
   padding-bottom: 10px;
 }
