+++ src/components/Feed/WeeklySummary/WeeklySummary.tsx
@@ -1,9 +1,10 @@
 import React from "react";
 import './WeeklySummary.scss';
 import toolsIcon from "../../../assets/icons/tools_icon.png";
-import { SheFeed, SheNestedStructure } from "../../../features/feed/she-feed.interface";
+import { SheFeed } from "../../../features/feed/she-feed.interface";
 import { WeeklySummaryList } from "./WeeklySummaryList";
 import weeklySummaryHeaderImage from '../../../assets/images/weekly_summary_header_image.png';
+import { transformWeeklySummary } from "./helper";
 
 type Props = {
     feedItem: SheFeed;
@@ -11,46 +12,8 @@ type Props = {
 
 export const WeeklySummary: React.FC<Props> = ({ feedItem }) => {
 
-  const transform = (structure?: { [key: string]: SheNestedStructure[] } )
-      : { source: string; content: string; badge: string; }[] | null => {
-    const weeklySummaryArray = [];
-    let hasSentiment = false;
-    let hasFitbit = false;
-    let contentSentiment = '';
-    let badgeSentiment = '';
-    let contentFitbit = '';
-    let badgeFitbit = '';
-
-    if (!structure) return null;
-
-    Object.keys(structure).map( key => {
-      if (key.includes('sentiment')) {
-        contentSentiment += structure[key][0].content + '\n';
-        badgeSentiment += structure[key][0].badge + '\n';
-        hasSentiment = true;
-      } else if (key.includes('fitbit')) {
-        contentFitbit += structure[key][0].content + '\n';
-        badgeFitbit += structure[key][0].badge + '\n';
-        hasFitbit = true;
-      } else {
-        weeklySummaryArray.push({ source: key, content: structure[key][0].content, badge: structure[key][0].badge });
-      }
-
-      return key;
-    });
-
-    if (hasSentiment) {
-      weeklySummaryArray.push({ source: 'sentiment', content: contentSentiment.trim(), badge: badgeSentiment });
-    }
-    if (hasFitbit) {
-      weeklySummaryArray.push({ source: 'fitbit', content: contentFitbit.trim(), badge: badgeFitbit });
-    }
-
-    return weeklySummaryArray;
-  };
-
   return (
-    <div className={'feed-item'}>
+    <div className="feed-item">
       <div className="weekly-summary-row">
         <div className="weekly-summary-header-container"
           style={{ backgroundImage: `url(${ weeklySummaryHeaderImage })` }}>
@@ -60,10 +23,10 @@ export const WeeklySummary: React.FC<Props> = ({ feedItem }) => {
           </div>
 
           <div className="weekly-summary-icon-container">
-            <img src={toolsIcon} alt="icon" width="26" height="26" />
+            <img src={toolsIcon} alt="Tool icon" width="26" height="26" />
           </div>
         </div>
-        <WeeklySummaryList nestedStructure={transform(feedItem.content?.nestedStructure)}/>
+        <WeeklySummaryList nestedStructure={transformWeeklySummary(feedItem.content?.nestedStructure)}/>
       </div>
     </div>
   );
